+++ docs/releasenotes/releasenotes.rst
@@ -17,6 +17,7 @@ Fixes
 * Fixes in autoswipe timeout computations
 * Dial: Slightly bigger default size of dial. Set block parameter ``scale: 0.9`` to reduce the dial size.
 * Graph: Remove total counter graph line for some dial types, only in case graphTypes and legend have not been defined.
+* Garbage: Fix for Circulus-Berkel
 
 V3.9.5 Beta (25-3-2022)
 -----------------------