+++ docs/releasenotes/releasenotes.rst
@@ -8,6 +8,9 @@ For Dashticz's **master** version Release Notes go to: https://dashticz.readthed
 Recent changes
 ---------------
 
+V3.9.6 Beta (18-10-2022)
+-----------------------
+
 Fixes
 ~~~~~
 
@@ -18,6 +21,7 @@ Fixes
 * Dial: Slightly bigger default size of dial. Set block parameter ``scale: 0.9`` to reduce the dial size.
 * Graph: Remove total counter graph line for some dial types, only in case graphTypes and legend have not been defined.
 * Garbage: Fix for Circulus-Berkel
+* Blinds: Fix Open/Close in new Domoticz version (build>14560)
 
 V3.9.5 Beta (25-3-2022)
 -----------------------